import_data/requirements.txt
absl-py==2.0.0
adafruit-io==2.7.0
antlr4-python3-runtime==4.9.2
asgiref==3.5.2
autopep8==2.0.0
cachetools==5.3.2
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==3.1.0
cryptography==41.0.5
distlib==0.3.6
Django==4.1.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.8.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
gspread==5.12.0
httplib2==0.22.0
idna==3.4
mysql-connector-python==8.0.31
mysqlclient==2.2.0
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
ortools==9.8.3296
paho-mqtt==1.6.1
pandas==2.1.3
pipenv==2022.11.11
platformdirs==2.5.4
protobuf==4.25.1
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==2.0.2
vercel==0.2.1
virtualenv==20.16.7
virtualenv-clone==0.5.7